updates as per PR feedback
