+++ blocks/products/src/editor.scss
@@ -4,14 +4,6 @@
  * Replace them with your own styles or remove the file completely.
  */
 
-// .wp-block-cata-shop {
-// 	color: #000;
-// }
-
-.wp-block-cata-products {
-	color: #000;
-}
-
 .wp-block-cata-products-fetch-btn {
 	margin-inline-end: 0.75em;
 }
